refactor: refactor details
